refactor: 모드 적용으로 위젯 합치기
